// src/shared/domain/chart/model/ChartColorOption.ts
export interface ChartColorOption {
  [key: string]: string;
  DR00: string;
  DR01: string;
  DR02: string;
  DR03: string;
  DR99: string;
  ES00: string;
  ES01: string;
  ES02: string;
  ES03: string;
  LR00: string;
  LR01: string;
  LR02: string;
  DC00: string;
  DC01: string;
  DC99: string;
  EC00: string;
  EC01: string;
  DG00: string;
  DG01: string;
  DG02: string;
  DG03: string;
  DG99: string;
  RR00: string;
  RR01: string;
  RR99: string;
}